Added more features to network
UUID of network is stored in a file so it's created only once. If UUID
changes network is undefined in libvirt before changes are applied.

Network can be also configured to use TFTP and custom DNS records can
be set. Domain is set to be localOnly keeping DNS requests on virtual
subnet local only, otherwise it causes loops.

Firewalld is configured to add virbr0 to trusted zone - only on RPM
based distros though.